requirements.txt
Django>=5.2,<6.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
drf-spectacular==0.29.0
psycopg2-binary==2.9.12
python-dotenv==1.2.2
openpyxl==3.1.5
gunicorn==23.0.0
